Add animation support for the Ripple, remove focusedItem

+++ src/controls/QskRadioBoxSkinlet.h
@@ -38,7 +38,7 @@ class QSK_EXPORT QskRadioBoxSkinlet : public QskSkinlet
         const QRectF&, QskAspect::Subcontrol, int index ) const override;
 
     QRectF textRect( const QskRadioBox*, const QRectF&, int ) const;
-    QRectF buttonRect( const QskRadioBox*, const QskAspect::Subcontrol target, const QRectF&, int ) const;
+    QRectF buttonRect( const QskRadioBox*, const QskAspect::Subcontrol target, const QRectF&, double ) const;
 
     QskAspect::States sampleStates( const QskSkinnable*,
         QskAspect::Subcontrol, int index ) const override;
